Static cast all implicit 64-to-32 bit int conversions in MgDriver
Need to clarify why the int64_t LogicalLocation.lxi and Mesh class
nrbxi loaded by the file are unlikely/cannot to exceed in32_t
limits

+++ src/multigrid/multigrid_driver.cpp
@@ -90,7 +90,8 @@ MultigridDriver::MultigridDriver(Mesh *pm, MGBoundaryFunc_t *MGBoundary, int inv
   }
   rootbuf_=new Real[pm->nbtotal*nvar_];
   mgtlist_ = new MultigridTaskList(this);
-  rootsrc_.NewAthenaArray(nvar_,pm->nrbx3,pm->nrbx2,pm->nrbx1);
+  rootsrc_.NewAthenaArray(nvar_, static_cast<int>(pm->nrbx3), static_cast<int>(pm->nrbx2),
+                          static_cast<int>(pm->nrbx1));
 }
 
 // destructor
@@ -215,9 +216,10 @@ void MultigridDriver::FillRootGridSource(void) {
     for (int n=0; n<pmy_mesh_->nbtotal; n++) {
       LogicalLocation &loc=pmy_mesh_->loclist[n];
       for (int v=0; v<nvar_; v++)
-        rootsrc_(v,loc.lx3,loc.lx2,loc.lx1)=rootbuf_[n*nvar_+v];
+        rootsrc_(v, static_cast<int>(loc.lx3), static_cast<int>(loc.lx2),
+                 static_cast<int>(loc.lx1))=rootbuf_[n*nvar_+v];
     }
-    mgroot_->LoadSource(rootsrc_,0,0,1.0);
+    mgroot_->LoadSource(rootsrc_, 0, 0, 1.0);
   }
   return;
 }
@@ -258,7 +260,8 @@ void MultigridDriver::TransferFromRootToBlocks(void) {
   } else {
     while(pmg!=NULL) {
       LogicalLocation &loc=pmg->loc_;
-      pmg->SetFromRootGrid(src, loc.lx1, loc.lx2, loc.lx3);
+      pmg->SetFromRootGrid(src, static_cast<int>(loc.lx1), static_cast<int>(loc.lx2),
+                           static_cast<int>(loc.lx3));
       pmg=pmg->next;
     }
   }
@@ -418,7 +421,9 @@ void MultigridDriver::SolveIterative(void) {
 
 void MultigridDriver::SolveCoarsestGrid(void) {
   Mesh *pm=pmy_mesh_;
-  int ni=std::max(pm->nrbx1, std::max(pm->nrbx2, pm->nrbx3)) >> (nrootlevel_-1);
+  int ni = (std::max(static_cast<int>(pm->nrbx1),
+                     std::max(static_cast<int>(pm->nrbx2), static_cast<int>(pm->nrbx3)))
+            >> (nrootlevel_-1));
   if (fperiodic_ && ni==1) { // trivial case - all zero
     mgroot_->ZeroClearData();
     return;
